JIT Engine + Test Refactoring
This commit refactors the JIT engine by:
- Inlining mlir_atomic_type_to_np_type into the new convert_mlir_atomic_type
- Misc: Adding more type hints, shortening variable names, etc.

This commit refactors the tests by combinging the two test files back into one larger test file with many test cases.
